% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{dual}
\alias{dual}
\title{Compute dual objective}
\usage{
dual(y, v, D, lambda)
}
\arguments{
\item{y}{response}

\item{v}{dual}

\item{D}{differencing matrix}

\item{lambda}{range variable}
}
\description{
Compute dual objective
}
