% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{kkt_residual}
\alias{kkt_residual}
\title{Compute KKT residual}
\usage{
kkt_residual(y, b, theta, v, D, lambda)
}
\arguments{
\item{y}{response}

\item{b}{primal variable}

\item{theta}{primal variable}

\item{v}{dual variable}

\item{D}{differencing matrix}

\item{lambda}{regularization parameter}
}
\description{
Compute KKT residual
}
